backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.2
python-dotenv>=1.0.1
SQLAlchemy>=2.0.36
psycopg[binary,pool]>=3.2.1
python-multipart>=0.0.6